Add annotation presentation service for api.read view
We will the calls to the presenter one-by-one into the new presentation
service. After that we have a central place where we can both present
the annotation with the presenter, and then merge any configured
formatters into that dictionary. So we display the same information
wherever we return an annotation back to the client.